// src/types/book.d.ts
// Generated by https://quicktype.io

export interface Books {
  book: Book;
}

export interface Book {
  book: any;
  title: string;
  pages: number;
  genre: string;
  cover: string;
  synopsis: string;
  year: number;
  ISBN: string;
  author: Author;
}

export interface Author {
  name: string;
  otherBooks: string[];
}

export type Genre = string;